Udating luni to Harmony r823222.
Highlights:
 - InputStream.skip concurrency issue
 - "better" messages in bound exceptions for streams and arrays
 - prefer fewer writes to underlying streams (using byte[] buffers)
 - Rename subclasses to not reuse names from their superclasses
 - PlatformAddressFactory.allocMap bugfix

Plus some spelling fixes, style fixes, serial version UIDs and other
boilerplate improvements.

+++ libcore/luni/src/main/native/org_apache_harmony_luni_util_fltparse.c
@@ -240,7 +240,7 @@ createFloat1 (JNIEnv * env, U_64 * f, IDATA length, jint e)
     }
   else if (e >= 0 && e < 39)
     {
-      result = (jfloat) (toDoubleHighPrecision (f, length) * pow (10.0, e));
+      result = (jfloat) (toDoubleHighPrecision (f, length) * pow (10.0, (double) e));
     }
   else if (e >= 39)
     {
@@ -260,7 +260,7 @@ createFloat1 (JNIEnv * env, U_64 * f, IDATA length, jint e)
       int dexp;
       U_32 fmant, fovfl;
       U_64 dmant;
-      dresult = toDoubleHighPrecision (f, length) / pow (10.0, -e);
+      dresult = toDoubleHighPrecision (f, length) / pow (10.0, (double) -e);
       if (IS_DENORMAL_DBL (dresult))
         {
           FLOAT_TO_INTBITS (result) = 0;
